tranformer/requirements.txt
absl-py==0.7.1
astor==0.7.1
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1
numpy==1.16.2
Pillow==6.0.0
pkg-resources==0.0.0
protobuf==3.7.1
pyrouge==0.1.3
six==1.12.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
torch==1.0.1.post2
torchvision==0.2.2.post3
Werkzeug==0.15.2

